requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
aiosqlite==0.20.0
python-dotenv==1.0.0
requests==2.31.0
openpyxl==3.1.2
cryptography==42.0.5
